Implemented Settings and UI work for enabling print guessing during deck list imports.
Added a section to the settings that contains default values for the print guessing.
Added an override setting to the deck parser classes that can be used to force-enable guessing for parsers that definitely require it (like the Magic Online parser).
Added Check boxes to the DeckImportWizard’s LoadListPage that take the default values from the settings and allow the user to enable both guessing and guessing preferences.

FossilOrigin-Name: 7c4f13da58199e19db04940e22166398fb8f0591820646eddc850e25af325efd