Implemented basic print guessing and hooked up the logic in the existing parsers.
If the name is given, normalize it (i.e. split it, if it is "Front face // Back face") and guess the language, if not set.
Then guess based on set code and collector number.
If that fails, guess based on the name.

FossilOrigin-Name: 92473b7bdfac7b882ebbb7f2712f68f209e5a3add7cbeb1f38bf34d3ffcb098c